Pythonサポートバージョンを3.9+に変更してmypy型エラーを修正
- Python 3.8サポートを削除（3.9+に変更）
- CI/CDのテスト対象をPython 3.9-3.12に限定
- mypy設定をPython 3.9に更新
- compound_interest.pyの戻り値を明示的にfloat型に変換
- cli.pyの関数引数にargparse.Namespace型注釈を追加
- mypy型チェックエラーを解消

+++ src/fintools/compound_interest.py
@@ -34,7 +34,7 @@ def annual_to_monthly_rate(annual_rate: float) -> float:
     if annual_rate < 0:
         raise ValueError("年利は0以上である必要があります")
 
-    return (1 + annual_rate) ** (1 / 12) - 1
+    return float((1 + annual_rate) ** (1 / 12) - 1)
 
 
 def monthly_to_annual_rate(monthly_rate: float) -> float:
